feat: Implement gender-specific calorie estimation formulas
This commit introduces the initial implementation for gender-specific calorie estimation formulas.

Changes made:
- Updated the `estimateCaloriesBurned` function in `lib/calorie-estimation.ts` to accept a `gender` parameter.
- Updated the `useCalorieTracker` hook to accept the user's gender and pass it to the `estimateCaloriesBurned` function.
- Added a gender selection UI to the `UserSettings` component.

Known issues:
- The calorie estimation formula for females is not yet correctly implemented, causing tests to fail. Further research is needed to find a scientifically validated formula.
